some updations

+++ src/app/input/input.component.ts
@@ -1,7 +1,8 @@
 import { HttpErrorResponse } from '@angular/common/http';
-import { Component, OnInit,ChangeDetectorRef, ChangeDetectionStrategy, Input,Output,EventEmitter } from '@angular/core';
+import { Component, OnInit,ChangeDetectorRef, ChangeDetectionStrategy, Input,Output,EventEmitter, ViewChild } from '@angular/core';
 import { Address } from 'src/app/Models/address';
 import { Customer } from 'src/app/Models/customer';
+import { AddressComponent } from '../address/address.component';
 import { AppServiceService } from '../app-service.service'
 import { State } from '../Models/BaseClass';
 import { Phone } from '../Models/phone';
@@ -17,6 +18,7 @@ export class InputComponent implements OnInit {
 
   private detector: ChangeDetectorRef;
   public confirmEmail: string='';
+  @ViewChild (AddressComponent) addForm!: AddressComponent;
 
   customerModel = new Customer(0,"","","","");
   AppService: AppServiceService;
@@ -28,6 +30,15 @@ export class InputComponent implements OnInit {
   }
   ngOnInit(): void {
     //this.getCustomerFromSearch();
+    
+  }
+
+  checkInputs()
+  {
+    if (!this.addForm) {
+      console.log(this.addForm)
+      console.log("child has missing inputs");
+    }
   }
 
   getCustomerFromSearch() {
@@ -45,30 +56,63 @@ export class InputComponent implements OnInit {
   }
 
   deleteAddress(address: Address){
+    if(address.addressId==0)
+    {
+      let value= this.customerModel.addresses.indexOf(address);
+      if(value> -1)
+      {
+        this.customerModel.addresses.splice(value,1);
+      }
+    }
     this.detector.detectChanges();
-    // this.customerModel.addresses=this.customerModel.addresses.filter(a => a !== address);
   }
 
   remainingAddresses(): Address[]{
     return this.customerModel.addresses.filter(a => a.stateEnum !== State.Deleted);
-    // return this.addressModel.phones.filter(a => a.stateEnum !== State.Deleted);
-
   }
 
   addAddress(status: boolean)
-  {
-    this.customerModel.addresses.push(new Address(0,'','','','','','',0,[
-      new Phone(0,"",0)
-    ])) 
+  { 
+    let phone: Phone= new Phone(0,'',0);
+    phone.stateEnum= State.Added;
+    let address: Address=new Address(0,'','','','','','',0,[phone]);
+    address.stateEnum= State.Added;
+    this.customerModel.addresses.push(address);
+    this.detector.detectChanges();
   }
 
   saveCustomer(){
+
+    if(!this.validate(this.customerModel)){
+      alert('Please provide complete details');
+      return;
+    }
+
+    this.AppService.saveCustomer(this.customerModel).subscribe(
+      (data: Customer) =>{
+        this.customerModel = {...data};
+        this.customerModel.confirmEmail=data.email;
+        this.resetStates();
+        alert("Data has been saved");
+        this.detector.detectChanges();
+        console.log(this.customerModel);
+      }, (error: HttpErrorResponse) =>{
+        console.log(error);
+      }
+    )
     this.AppService.saveCustomer(this.customerModel);
   }
 
   deleteCustomer(){
     this.customerModel.stateEnum= State.Deleted;
-    this.AppService.deleteCustomer(this.customerModel)
+    if(this.customerModel.customerId==0)
+    {
+      alert("Error! Customer has not been created yet");
+      return;
+    }
+    this.AppService.deleteCustomer(this.customerModel);
+    alert("Customer is deleted");
+    this.customerModel= new Customer(0,"","","","");
   }
 
   changeState()
@@ -76,9 +120,53 @@ export class InputComponent implements OnInit {
     this.customerModel.stateEnum= State.Modified;
   }
 
-  changeParentState(value: boolean)
+  resetStates()
   {
-    this.customerModel.stateEnum= State.Modified;
-
+    this.customerModel.stateEnum=State.Unchanged;
+    for(let add of this.customerModel.addresses){
+      add.stateEnum= State.Unchanged;
+      for(let ph  of add.phones)
+        ph.stateEnum= State.Unchanged;
+    }
   }
+
+  validate(object: any) : boolean {
+    for(let key in object){
+      let property = object[key];
+      if(typeof property === 'string'){
+        if(property.length == 0){
+          return false;
+        }
+      }
+      else if(Array.isArray(property)){
+        for(let element of property){
+          return this.validate(element);
+        }
+      }
+    }
+    return true;
+  }
+  
+//   validate(customer: Customer){
+//     for(let address of customer.addresses)
+//     {
+//       if(address.addressType == '')
+//       {
+//         alert('invalid street1');
+//         return;
+//       }
+//       if(address.street1 == '')
+//       {
+//         alert('invalid street1');
+//         return;
+//       }
+//       if(address.zipcode == '')
+//       {
+//         alert('invalid street1');
+//         return;
+//       }
+      
+//     }
+// }
+
 }
